refactored setup and haunt to abide by style guide
